Extract `TestStringUtil.lineRanges`
